sdk: drop bigIntReplacer; wire types are now plain JSON numbers

+++ sdk/src/codegen.ts
@@ -519,7 +519,7 @@ function findAttachPair(
   const attach = methods.find(([n]) => n === "attach")?.[1].function;
   const detach = methods.find(([n]) => n === "detach")?.[1].function;
   if (attach == null || detach == null) return null;
-  if (!attach.params.some((p) => p.name === "vout" && p.type === "u64")) {
+  if (!attach.params.some((p) => p.name === "vout" && p.type === "u32")) {
     return null;
   }
   if (!isResultRef(attach.result, ctx) || !isResultRef(detach.result, ctx)) {
